move `DragController` out from `InventoryView`
