+++ Packages/com.fullmetalbagel.dope-inventory-ugui/Runtime/UI/InventoryView.DragController.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: e43af904e1cc4f90a256886ab5b19052
+timeCreated: 1758690607
